[Pre-Release] Prepare for release DSS-1.1.0

+++ dss-framework/dss-framework-workspace-server/src/main/java/com/webank/wedatasphere/dss/framework/workspace/service/DSSSideInfoService.java
@@ -24,5 +24,5 @@ import java.util.List;
 
 public interface DSSSideInfoService {
 
-    List<SidebarVO> getSidebarVOList(String username, Integer workspaceId,boolean isEnglish);
+    List<SidebarVO> getSidebarVOList(String username, Long workspaceId,boolean isEnglish);
 }
